Removed all over casting

+++ dvt/samples/keep_alive_test/firmware.c
@@ -26,8 +26,8 @@ typedef struct {
 /* list of all supported firmware target info */
 static firmware_list_t firmware_list[] = {
     /* version     code_size     name_spec          description */
-    {0x01000000, (uint32_t)-1, ".*\\.[tT][xX][tT]", "Text file"},     /* any *.txt files */
-    {0x00000100, (uint32_t)-1, ".*\\.[bB][iI][nN]", "Binary Image" }  /* any *.bin files */
+    {0x01000000, -1, ".*\\.[tT][xX][tT]", "Text file"},     /* any *.txt files */
+    {0x00000100, -1, ".*\\.[bB][iI][nN]", "Binary Image" }  /* any *.bin files */
 };
 
 static int firmware_download_started = 0;
